Feat/socket

+++ src/board-messages/board-messages.service.ts
@@ -1,8 +1,6 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import { HttpException, HttpStatus, Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Board_Message } from 'src/_common/entities/board-message.entity';
-import { BoardMembersService } from 'src/board-members/board-members.service';
-import { BoardsService } from 'src/boards/boards.service';
 import { UsersService } from 'src/users/users.service';
 import { Repository } from 'typeorm';
 
@@ -11,16 +9,14 @@ export class BoardMessagesService {
   constructor(
     @InjectRepository(Board_Message)
     private boardMessageRepository: Repository<Board_Message>,
-    private boardsService: BoardsService,
-    private usersService: UsersService,
-    private boardMembersService: BoardMembersService
+    private usersService: UsersService
   ) {}
 
   //보드 메세지 조회
   async GetBoardMessages(joinBoards: any): Promise<any> {
     return Promise.all(
-      joinBoards.map(async (board: any) => {
-        const messageInfos = await this.boardMessageRepository
+      joinBoards.map((board: any) => {
+        return this.boardMessageRepository
           .createQueryBuilder('message')
           .innerJoinAndSelect('message.user', 'user')
           .innerJoinAndSelect('message.board', 'board')
@@ -41,7 +37,6 @@ export class BoardMessagesService {
           .where('message.board_id = :boardId ', { boardId: board.board_id })
           .orderBy('message.created_at')
           .getRawMany();
-        return messageInfos;
       })
     );
   }
@@ -60,8 +55,11 @@ export class BoardMessagesService {
     });
   }
 
-  async deleteMessage(messageId: number) {
-    return await this.boardMessageRepository.delete({ id: messageId });
+  async deleteMessage(messageId: number): Promise<void> {
+    const existMessage = await this.boardMessageRepository.findOne({ where: { id: messageId } });
+    if (!existMessage) throw new HttpException({ message: '해당 메시지를 찾을 수 없습니다.' }, HttpStatus.NOT_FOUND);
+
+    await this.boardMessageRepository.delete({ id: messageId });
   }
 
   //보드 메세지 멘션 추출
